requirements.txt
xlwt==1.0.0
djangorestframework-jwt==1.7.2
Babel==1.3
Django==1.8.4
mysql-python==1.2.5
Pillow==2.7.0
Unidecode==0.04.18
argparse==1.2.1
cffi==1.2.1
cryptography==1.0.2
celery==3.1.19
decorator==4.0.2
django-appconf==1.0.1
django-celery==3.1.17
django-celery-with-redis==3.0
django-ckeditor==5.0.2
django-compressor==1.5
django-extra-views==0.6.4
django-haystack==2.3.1
django-oscar==1.1.1
django-simple-captcha==0.4.6
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.3
djangorestframework==3.2.4
enum34==1.0.4
factory-boy==2.4.1
httplib2==0.9.2
idna==2.0
ipaddress==1.0.14
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
mock==1.0.1
ndg-httpsclient==0.4.0
path.py==8.1.1
pexpect==3.3
phonenumbers==6.3.0
pickleshare==0.5
psycopg2==2.6.1
purl==1.1
pusher==1.2.3
pyOpenSSL==0.15.1
pyasn1==0.1.9
pycountry==1.15
pycparser==2.14
pycrypto==2.6.1
pysolr==3.3.2
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
rsa==3.3
simplegeneric==0.8.1
six==1.9.0
sorl-thumbnail==12.2
traitlets==4.0.0
urllib3==1.12
wsgiref==0.1.2
